+++ NEWS.md
@@ -1,4 +1,10 @@
-# colt 1.0.0
+# colt 0.1.0.9000
+
+* colt now guesses the console color theme (via **rstudioapi**) and sets the 
+  appropriate default theme.
+
+
+# colt 0.1.0
 
 * Initial release.